# src/main/java/com/ejemplo/demo/controller/PruebaResponseController.java
package com.ejemplo.demo.controller;

import java.util.Map;



import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;

import com.ejemplo.demo.dto.EstadoResponse;
import com.ejemplo.demo.dto.SaludoRequest;
import com.ejemplo.demo.dto.SaludoResponse;
import com.ejemplo.demo.service.SaludoService;

import jakarta.validation.Valid;

@RestController
@RequestMapping("/api/v1")
public class PruebaResponseController {

	@GetMapping
	public ResponseEntity<EstadoResponse> respuesta() {
		
		EstadoResponse respuesta = new EstadoResponse(
				"Ok",
				"Workshop Springbook activo");
		
		
		return ResponseEntity.ok(respuesta);
	}
	
	private final SaludoService saludoService;
	
	public PruebaResponseController(SaludoService saludoService) {
		this.saludoService = saludoService;
	}
	
	@GetMapping("/saludos")
	public ResponseEntity<SaludoResponse> saludar(
			@RequestParam(defaultValue = "Mundo") String nombre){
		return ResponseEntity.ok(saludoService.crearSaludo(nombre));
	}
	
	@PostMapping("/saludos")
	public ResponseEntity<SaludoResponse> saludarPost(
			@Valid @RequestBody SaludoRequest request){
		return ResponseEntity.ok(saludoService.crearSaludo(request.nombre()));
	}
	
	
	
	
	
//	@GetMapping("/error")
//	public ResponseEntity<String> error() {
//		return ResponseEntity.badRequest().body("no funciono");
//	}
}