# src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Demo API Spring Boot
  version: 1.0.0
...
paths:
...
  /api/v1/saludos:
    get:
      tags:
        - workshop
      operationId: saludarPorGet
      parameters:
        - name: nombre
          in: query
          required: false
          schema:
            type: string
            default: Mundo
      responses:
        "200":
          description: Saludo por GET
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaludoResponse"

    post:
      tags:
        - workshop
      operationId: saludarPorPost
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaludoRequest"
      responses:
        "200":
          description: Saludo por POST
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaludoResponse"
...
components:
  schemas:
...
    SaludoRequest:
      type: object
      properties:
        nombre:
          type: string
    SaludoResponse:
      type: object
      properties:
        mensaje:
          type: string
        timestamp:
          type: string
          format: date-time
...